Network-side rewrite and remove Android stuffs
- Network-side rewrite
  - From now on, this sdk differentiates clientbound and serverbound packets and packages
- Remove Android-related imports
- [TEMPORARY] Create `ConcurrentTask` class for seamless translation from `AsyncTask`
- Move `DeviceManager` to the root package
- Prepare multi device type support by separating AC specific features to separate classes.
