+++ gradle.properties
@@ -1,5 +1,5 @@
 artifactId = greesdk
 name = GreeSDK
-version = 0.1.0
+version = 1.0.0
 
 mavenUrl = https://maven.czsoft.eu/
