+++ src/main/java/eu/czsoft/greesdk/DeviceManagerEvent.java
@@ -1,0 +1,6 @@
+package eu.czsoft.greesdk;
+
+public enum DeviceManagerEvent {
+    DEVICE_LIST_UPDATED,
+    DEVICE_STATUS_UPDATED
+}
