+++ src/main/java/eu/czsoft/greesdk/DeviceManagerEventListener.java
@@ -1,0 +1,6 @@
+package eu.czsoft.greesdk;
+
+public interface DeviceManagerEventListener {
+
+    void onEvent(DeviceManagerEvent event);
+}
