+++ src/main/java/eu/czsoft/greesdk/Unzipped.java
@@ -1,0 +1,4 @@
+package eu.czsoft.greesdk;
+
+public record Unzipped(String[] keys, Integer[] values) {
+}
