+++ src/main/java/eu/czsoft/greesdk/appliances/airconditioner/FanSpeed.java
@@ -1,0 +1,10 @@
+package eu.czsoft.greesdk.appliances.airconditioner;
+
+public enum FanSpeed {
+    AUTO,
+    LOW,
+    MEDIUM_LOW,
+    MEDIUM,
+    MEDIUM_HIGH,
+    HIGH
+}
