+++ src/main/java/eu/czsoft/greesdk/appliances/airconditioner/Mode.java
@@ -1,0 +1,10 @@
+package eu.czsoft.greesdk.appliances.airconditioner;
+
+public enum Mode {
+    AUTO,
+    COOL,
+    DRY,
+    FAN,
+    HEAT
+}
+
