+++ src/main/java/eu/czsoft/greesdk/appliances/airconditioner/TemperatureUnit.java
@@ -1,0 +1,6 @@
+package eu.czsoft.greesdk.appliances.airconditioner;
+
+public enum TemperatureUnit {
+    CELSIUS,
+    FAHRENHEIT
+}
