+++ src/main/java/eu/czsoft/greesdk/appliances/airconditioner/VerticalSwingMode.java
@@ -1,0 +1,16 @@
+package eu.czsoft.greesdk.appliances.airconditioner;
+
+public enum VerticalSwingMode {
+    DEFAULT,
+    FULL,
+    FIXED_TOP,
+    FIXED_MIDDLE_TOP,
+    FIXED_MIDDLE,
+    FIXED_MIDDLE_BOTTOM,
+    FIXED_BOTTOM,
+    SWING_BOTTOM,
+    SWING_MIDDLE_BOTTOM,
+    SWING_MIDDLE,
+    SWING_MIDDLE_TOP,
+    SWING_TOP
+}
