--- src/main/java/eu/czsoft/greesdk/device/DeviceManagerEventListener.java
@@ -1,10 +1,0 @@
-package eu.czsoft.greesdk.device;
-
-public interface DeviceManagerEventListener {
-    enum Event {
-        DEVICE_LIST_UPDATED,
-        DEVICE_STATUS_UPDATED
-    }
-
-    void onEvent(Event event);
-}
