+++ src/main/java/eu/czsoft/greesdk/net/communication/AsyncCommunicationFinishedListener.java
@@ -1,0 +1,5 @@
+package eu.czsoft.greesdk.net.communication;
+
+public abstract class AsyncCommunicationFinishedListener {
+    public abstract void onFinished();
+}
