+++ src/main/java/eu/czsoft/greesdk/net/packets/ClientboundPacket.java
@@ -1,0 +1,12 @@
+package eu.czsoft.greesdk.net.packets;
+
+import lombok.ToString;
+
+/**
+ * <b>S</b>erver-><b>C</b>lient<br/>
+ * ClientboundPacket means, that it is sent from the <i>Server</i> to the <i>Client</i>, so in this case, from Gree appliance to this Sdk.
+ */
+@ToString(callSuper = true)
+public class ClientboundPacket extends Packet {
+
+}
