+++ src/main/java/eu/czsoft/greesdk/net/packets/Packet.java
@@ -1,0 +1,30 @@
+package eu.czsoft.greesdk.net.packets;
+
+import com.google.gson.annotations.SerializedName;
+import eu.czsoft.greesdk.net.packets.packs.Pack;
+import lombok.ToString;
+
+/**
+ * In this documentation, the Sdk is defined as Client and Gree appliances are the Server.
+ * This interface ties both types of packages together.
+ */
+@ToString(exclude = {"encryptedPack"})
+public class Packet {
+
+    @SerializedName("t")
+    public String type;
+
+    @SerializedName("tcid")
+    public String targetClientId;
+    @SerializedName("i")
+    public Integer i;
+    @SerializedName("uid")
+    public Integer uid;
+    @SerializedName("cid")
+    public String clientId;
+
+    @SerializedName("pack")
+    public String encryptedPack;
+
+    public transient Pack pack;
+}
