+++ src/main/java/eu/czsoft/greesdk/net/packets/clientbound/ResponsePacket.java
@@ -1,0 +1,6 @@
+package eu.czsoft.greesdk.net.packets.clientbound;
+
+import eu.czsoft.greesdk.net.packets.ClientboundPacket;
+
+public class ResponsePacket extends ClientboundPacket {
+}
