--- src/main/java/eu/czsoft/greesdk/network/AsyncCommunicationFinishedListener.java
@@ -1,5 +1,0 @@
-package eu.czsoft.greesdk.network;
-
-public abstract class AsyncCommunicationFinishedListener {
-    public abstract void onFinished();
-}
