--- src/main/java/eu/czsoft/greesdk/packets/ScanPacket.java
@@ -1,9 +1,0 @@
-package eu.czsoft.greesdk.packets;
-
-public class ScanPacket extends Packet {
-    public static String TYPE = "scan";
-
-    public ScanPacket() {
-        type = TYPE;
-    }
-}
