--- src/main/java/eu/czsoft/greesdk/packs/BindOkPack.java
@@ -1,16 +1,0 @@
-package eu.czsoft.greesdk.packs;
-
-import com.google.gson.annotations.SerializedName;
-
-public class BindOkPack extends Pack {
-    public static String TYPE = "bindok";
-
-    public String key;
-
-    @SerializedName("r")
-    public int resultCode;
-
-    public BindOkPack() {
-        type = TYPE;
-    }
-}
