--- src/main/java/eu/czsoft/greesdk/packs/BindPack.java
@@ -1,11 +1,0 @@
-package eu.czsoft.greesdk.packs;
-
-public class BindPack extends Pack {
-    public static String TYPE = "bind";
-
-    public int uid;
-
-    public BindPack() {
-        type = TYPE;
-    }
-}
