--- src/main/java/eu/czsoft/greesdk/packs/DatPack.java
@@ -1,20 +1,0 @@
-package eu.czsoft.greesdk.packs;
-
-import com.google.gson.annotations.SerializedName;
-
-public class DatPack extends Pack {
-    public static String TYPE = "dat";
-
-    @SerializedName("r")
-    public int resultCode;
-
-    @SerializedName("cols")
-    public String[] keys;
-
-    @SerializedName("dat")
-    public Integer[] values;
-
-    public DatPack() {
-        type = TYPE;
-    }
-}
