--- src/main/java/eu/czsoft/greesdk/packs/Pack.java
@@ -1,10 +1,0 @@
-package eu.czsoft.greesdk.packs;
-
-import com.google.gson.annotations.SerializedName;
-
-public class Pack {
-    @SerializedName("t")
-    public String type;
-
-    public String mac;
-}
